pyproject.toml
[build-system]
requires = ["setuptools>=66.1"]
build-backend = "setuptools.build_meta:__legacy__"

[project]
name = "clusterx"
authors = [{name = "Haochen Ye"}, {name = "HaiJun Lv"}, {name = "HaoDong Chen"}]
maintainers = [{name = "Haochen Ye"}]
requires-python = ">=3.10"
dynamic = ["version", "readme"]
dependencies = [
    "prompt_toolkit",
    "toml",
    "cyclopts",
    "rich",
    "pydantic>=2.10.6",
    "more_itertools",
    "pyyaml",
    "aiohttp",
    "requests",
    "tenacity"
]

[project.optional-dependencies]
volc = ["volcengine"]
test = [
    "pytest",
]

[tool.setuptools.dynamic]
readme = {file = "README.md", content-type = "text/markdown"}

[tool.setuptools.packages.find]
where = ["."]
include = ["clusterx*"]

[[tool.mypy.overrides]]
module = ["requests.*", "toml.*", "yaml.*"]
ignore_missing_imports = true

[project.scripts]
clusterx = "clusterx.cli:app"

[tool.black]
line-length = 120

[tool.pyright]
reportRedeclaration = "none"

[tool.flake8]
max-line-length = 120

[tool.isort]
profile = "black"

[tool.mypy]
ignore_missing_imports = true
